refactor: remove repeated typization and get rid of interfaces
