+++ src/api/apiClient.ts
@@ -1,4 +1,4 @@
-import type { SearchResponse } from '../types/interfaces';
+import type { SearchResponse } from '../types/types';
 
 export const searchRequest = async (title: string): Promise<SearchResponse> => {
   const result = await fetch('https://stapi.co/api/v1/rest/animal/search', {
